backend/pyproject.toml
[project]
name = "loops-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "fsrs>=6.3.0",
    "google-genai>=1.51.0",
    "greenlet>=3.2.4",
    "openai>=1.50.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "supabase>=2.15.0",
    "uvicorn[standard]>=0.38.0",
    "nltk>=3.9.2",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "langgraph>=1.0.5",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.3",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "freezegun>=1.5.0",
    "factory-boy>=3.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "E712", # SQLModel/SQLAlchemy requires == True for boolean comparisons
    "E711", # SQLModel/SQLAlchemy requires == None for null comparisons
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # unused imports

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/app/services --cov-report=term-missing --cov-fail-under=60"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
markers = ["e2e: marks tests as end-to-end tests that use real external APIs"]

[tool.coverage.run]
omit = [
    "src/app/services/word_tutor_graph.py", # LangGraph workflow requires complex mocking
]
